agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.1",
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.4.3",
    "numpy>=2.4.2",
    "piper-tts>=1.4.1",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
]
